# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Boleteria Online",
    "description": "Colleción de API's correspondientes al sistema de Pasajes Online.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@boleteriasystem.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseICollectionWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseICollectionWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseICollectionWebResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClienteRequest": {
        "required": [
          "dni",
          "fecha_nacimiento",
          "genero",
          "nacionalidad",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "dni": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "fecha_nacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "nacionalidad": {
            "type": "string"
          },
          "genero": {
            "$ref": "#/components/schemas/Gender"
          }
        },
        "additionalProperties": false
      },
      "ClienteResponse": {
        "required": [
          "dni",
          "fecha_nacimiento",
          "genero",
          "nacionalidad",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "dni": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "fecha_nacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "nacionalidad": {
            "type": "string"
          },
          "genero": {
            "$ref": "#/components/schemas/Gender"
          }
        },
        "additionalProperties": false
      },
      "ClienteResponseICollectionWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClienteResponse"
            },
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClienteResponseWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ClienteResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Gender": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}